chore: renaming db dir, updating favicon, re-arranging comments, housekeeping
